GunDetectedPicture
Here is a screenshot of the program in action, as you can see I am holding a picture of a gun up to the camera and it is able to identity it as a gun. Keep in mind that this picture of a gun wasn't used when training this model, this is a picture of a gun this model has never seem before.